+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- BREAKING FEATURE: modify case specification to contain `pre_once`, `pre_case`, `post_case`, and `post_once` fields in `hooks` table (instead of previous `pre` and `post`). This allows for a more granular control over case execution.
+
 # Version 0.13.0.9000
 
 - Stop official support of Neovim 0.7.